pyproject.toml
[project]
name = "sim"
version = "0.1.0"
description = ""
authors = [
    {name = "Augustine77",email = "hugwolf77@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.11.11"
dependencies = [
    "jupyter (>=1.1.1,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "numpy (==1.26)",
    "scipy (>=1.15.2,<2.0.0)",
    "simpy (>=4.1.1,<5.0.0)",
    "torch (>=2.5.1,<3.0.0)",
    "torchvision (>=0.20.1,<0.21.0)",
    "torchaudio (>=2.5.1,<3.0.0)",
    "torchmetrics (>=1.7.1,<2.0.0)",
    "torchinfo (>=1.8.0,<2.0.0)",
    "torchsummary (>=1.5.1,<2.0.0)",
    "tensorflow (==2.17.0)",
    "tensorboard (==2.17.0)",
    "mlflow (>=2.21.3,<3.0.0)",
    "keras (>=3.9.2,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "shap (>=0.47.1,<0.48.0)",
    "orange3 (>=3.38.1,<4.0.0)",
    "pyqt5 (>=5.15.11,<6.0.0)",
    "pyqtwebengine (>=5.15.7,<6.0.0)",
    "pyside6 (>=6.8.2.1,<7.0.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "pyshark (>=0.6,<0.7)",
    "pygame (>=2.6.1,<3.0.0)",
    "mcp[cli] (>=1.6.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch"}
torchvision = {source = "pytorch"}
torchaudio = {source = "pytorch"}
